Make order pickers complete order process
Order pickers will now go to a specified bin,
then go to the shipping bin, then home. If a
path is blocked, the picker will yield. If a
picker continues in the yield state, it will
home in order to clear the warehouse floor,
then attempt to retrieve or ship an item,
depending on the context. The compiled
warehouse program has a small demo using
4 order pickers.